+++ src/fractalgebra/helpers/__init__.py
@@ -1,0 +1,4 @@
+from .calculator import Calc
+from .errors import InvalidFractionError, InvalidInputError
+from .main_helper import Fractalgebra
+from .rational_number import RationalNumber
